backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
drf-spectacular>=0.26
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
gunicorn>=21.2
pytest>=8.0
pytest-django>=4.8
python-dotenv>=1.0
Pillow>=10.0
dj-database-url>=2.1.0
